GameLogic, Microphone component made

+++ Engine/CinderEngine/include/Engine/System/Physics/Resolution.h
@@ -21,15 +21,15 @@ namespace Framework
 		static const int maxContacts = 256;
 		unsigned numOfContacts;
 
-		void ResolvePositions(const double dt);
-		void ResolveVelocities(const double dt);
+		void ResolvePositions(const float dt);
+		void ResolveVelocities(const float dt);
 
 	public:
 		ContactSet(){};
 		//BodyContact * getNextContact();
 		BodyContact contactArray[maxContacts];
 		void addContact(BodyContact * contact);
-		void ResolveContacts(const double dt);/*
+		void ResolveContacts(const float dt);/*
 		void ResolveStatic(CollisionEvent* pre, CollisionEvent* post,
 			RigidBody::DynamicState stateA, RigidBody::DynamicState stateB);
 		void ResolveDynamic(CollisionEvent* pre, CollisionEvent* post);*/
